Transports/Seatrac: Extract driver.
